# Replace (indptr, indices, data) tuple with CSGraph jitclass