--- a/benchmarks/bench_skan.py
+++ b/benchmarks/bench_skan.py
@@ -22,7 +22,7 @@ def timer():
 
 def bench_suite():
     times = OrderedDict()
-    skeleton = np.load('infected3.npz')['skeleton']
+    skeleton = np.load(os.path.join(rundir, 'infected3.npz'))['skeleton']
     with timer() as t_build_graph:
         g, indices, degrees = csr.skeleton_to_csgraph(skeleton,
                                                       spacing=2.24826)