--- a/requirements.txt
+++ b/requirements.txt
@@ -2,3 +2,4 @@ numpy>=1.11
 scipy>=1.18
 networkx>=1.11
 numba>=0.29
+pandas>=0.18