--- a/setup.cfg
+++ b/setup.cfg
@@ -1,2 +1,5 @@
 [tool:pytest]
-addopts = --doctest-modules --cov-report term-missing --cov . --ignore setup.py
+addopts = --doctest-modules  --ignore setup.py --ignore benchmarks/
+; use "--cov-report term-missing --cov ." to also run coverage.
+; remember to clear numba cache and disable Numba JIT for accurate coverage.
+; use "make numba-clean" and "export NUMBA_DISABLE_JIT=1".