--- a/test-environment.yml
+++ b/test-environment.yml
@@ -2,6 +2,7 @@ name: skantest
 dependencies:
 - python>=3.5*
 - setuptools>=19.6*
+- numba>=0.29
 - numpy>=1.11*
 - numpydoc>=0.5*
 - scipy>=0.17*