This is [cannot currently be combined](https://groups.google.com/a/continuum.io/forum/#!msg/numba-users/6-AFAJgRF6s/q1CqDSBoAAAJ) with cache=True for functions that read in the class, but the warmup time is acceptable:

Timings before this PR:
---  build graph 0.302 s
---  compute statistics 0.212 s
---  compute per-skeleton statistics 0.469 s

Timings with this PR:
---  build graph 0.298 s
---  compute statistics 1.037 s
---  compute statistics again 0.190 s
---  compute per-skeleton statistics 0.456 s